Several Purges of Toasts and Unused mobile elements purged in mobile And Unit Test is finalized now.
Several redundant Toasts and unused elements are removed in those pages.
- CommunitiesPageActivity
- CreatePostPage
- EditProfile
- HomePageActivity
- InsidePost
- LoginPageActivity
- SignUpPAgeActivity

And also Unittest of Automated UI test is now contains:
- test_goLoginPage
- test_login
- test_goMyProfilePage
- test_login_logout
- test_open_my_communities
- test_open_all_communities
- test_filter_all_communities
- test_open_post
- test_filter_post

[Ticket: Muhammet Çavuş]
