+++ mobile/FaceSpace/app/src/main/java/com/example/facespace/CommunitiesPageActivity.kt
@@ -121,7 +121,7 @@ class CommunitiesPageActivity : AppCompatActivity() {
                 }
             },
             Response.ErrorListener { error ->
-                Toast.makeText(this, params.toString()+error.toString(), Toast.LENGTH_SHORT).show()
+                Toast.makeText(this, error.toString(), Toast.LENGTH_SHORT).show()
             }) {
             override fun getParams(): Map<String, String> {
                 return params
