+++ mobile/UIUnitTest/.idea/misc.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.8 (UIUnitTest)" project-jdk-type="Python SDK" />
+</project>
